Refactors centroiding
